+++ dart/geodata/example/geodata_example.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2020-2023 Navibyte (https://navibyte.com). All rights reserved.
+// Copyright (c) 2020-2024 Navibyte (https://navibyte.com). All rights reserved.
 // Use of this source code is governed by a “BSD-3-Clause”-style license that is
 // specified in the LICENSE file.
 //
@@ -12,6 +12,9 @@ To test run this from command line: dart example/geodata_example.dart
 
 import 'package:geobase/geobase.dart';
 import 'package:geodata/ogcapi_features_client.dart';
+
+import 'package:http/http.dart' as http;
+import 'package:http/retry.dart';
 
 /// This example demonstrates accessing metadata and geospatial feature items
 /// from a GeoJSON based feature collection provided by a RESTful service
@@ -36,6 +39,11 @@ Future<void> main(List<String> args) async {
   // create an OGC API Features client for the open pygeoapi demo service
   // (see https://pygeoapi.io/ and https://demo.pygeoapi.io for more info)
   final client = OGCAPIFeatures.http(
+    // set the HTTP client using the standard HTTP retry client on API calls
+    // (when not set, the default `http.Client()` is used without retry logic)
+    client: RetryClient(http.Client(), retries: 4),
+
+    // The API endpoint serving an OGC API Features demo service (of pygeoapi)
     endpoint: Uri.parse('https://demo.pygeoapi.io/master/'),
   );